+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 ## master
 
 ## 1.1.5RC1
+* Abstract do_validate_settings for easier reuse
 * Don't shortcode attribute for reset options
 * Move get_suggest to widget abstract class from interface
+* RESOLVE Add widget option defaults
 * RESOLVE michael-cannon/testimonials-widget#6 On widget options, switch labels expand/collapse when using them
 * RESOLVE michael-cannon/testimonials-widget#78 Show revert link in deactivation admin notice
+* RESOLVE Move form_instance to abstract class Aihrus_Widget
+* RESOLVE Move get_suggest to abstract class Aihrus_Widget
+* RESOLVE Remove tw- reference
+* RESOLVE Use widget args than globals
 * Update licensing GPL 2.0
 
 ## 1.1.4